// src/app/core/models/configuracion-personalizada-dto.ts
export interface ConfiguracionPersonalizadaDto {
    idConfiguracionPersonalizada?: number;
    codigoConfiguracion?: string;
    userName?: string;
    codigoTransaccion?: string;
    idSucursal?: number;
    referencia1?: string;
    referencia2?: string;
    valor?: string;
    activo?: boolean;
    usuarioCreacion?: string;
    fechaCreacion?: string;
    equipoCreacion?: string;
    usuarioModificacion?: string;
    fechaModificacion?: string;
    equipoModificacion?: string;
}